Implement fine grained reactivity for items
If an operation does not affect the number of items in ListItems
then just update the item that was changed. This is done by ensuring
that the object that is passed to setItem is not the same object as the
signal. Mutating the object will not trigger a re-render.

If an operation does affect the number of items in ListItems then
rerender the whole list by calling toggleRefreshData().